fix #232101 Inability to create new text lines in frames
