+++ libmscore/undo.cpp
@@ -2188,73 +2188,6 @@ void ChangeGap::flip(EditData*)
       }
 
 //---------------------------------------------------------
-//  ChangeText::insertText
-//---------------------------------------------------------
-
-void ChangeText::insertText(EditData* ed)
-      {
-      TextCursor tc = c;
-      c.text()->editInsertText(&tc, s);
-      c.text()->triggerLayout();
-      if (ed) {
-            TextCursor* ttc = c.text()->cursor(*ed);
-            *ttc = tc;
-            }
-      }
-
-//---------------------------------------------------------
-//  ChangeText::removeText
-//---------------------------------------------------------
-
-void ChangeText::removeText(EditData* ed)
-      {
-      TextCursor tc = c;
-      TextBlock& l  = c.curLine();
-      int column    = c.column();
-
-      for (int n = 0; n < s.size(); ++n)
-            l.remove(column);
-      c.text()->triggerLayout();
-      if (ed)
-            *c.text()->cursor(*ed) = tc;
-      }
-
-//---------------------------------------------------------
-//   SplitText
-//---------------------------------------------------------
-
-void SplitText::undo(EditData* ed)
-      {
-      TextCursor tc = c;
-      tc.deleteChar();
-      c.text()->triggerLayout();
-      if (ed)
-            *c.text()->cursor(*ed) = tc;
-      }
-
-void SplitText::redo(EditData* ed)
-      {
-      TextCursor tc = c;
-      TextBase* t   = tc.text();
-      int line      = tc.row();
-
-      CharFormat* charFmt = tc.format();         // take current format
-      t->textBlockList().insert(line + 1, tc.curLine().split(tc.column()));
-      t->textBlock(line).setEol(true);
-      t->setTextInvalid();
-      if (t->textBlockList().last() != t->textBlock(line+1))
-            t->textBlock(line+1).setEol(true);
-
-      c.text()->triggerLayout();
-      if (ed) {
-            tc.setRow(line+1);
-            tc.setColumn(0);
-            tc.setFormat(*charFmt);             // restore orig. format at new line
-            *c.text()->cursor(*ed) = tc;
-            }
-      }
-
-//---------------------------------------------------------
 //   JoinText
 //---------------------------------------------------------
